:recycle: api: move getImportTemplate from session controller to certif center controller
