+++ api/tests/unit/application/session/index_test.js
@@ -1027,18 +1027,4 @@ describe('Unit | Application | Sessions | Routes', function () {
       expect(response.statusCode).to.equal(200);
     });
   });
-
-  describe('GET /api/sessions/import', function () {
-    it('should exist', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/sessions/import');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-  });
 });
