+++ certif/app/components/sessions/panel-header.js
@@ -17,7 +17,7 @@ export default class PanelHeader extends Component {
 
   @action
   async downloadSessionImportTemplate() {
-    const url = '/api/sessions/import';
+    const url = '/api/certification-centers/import';
     const token = this.session.data.authenticated.access_token;
     try {
       await this.fileSaver.save({ url, token });
